# openapi/el-trigal-api.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "El Trigal API",
    "version": "1.0.0",
    "description": "API REST independiente para el website Astro y el panel Angular."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/public/faqs": {
      "get": {
        "tags": ["public/content"],
        "operationId": "listPublicFaqs",
        "parameters": [
          { "$ref": "#/components/parameters/locale" },
          { "$ref": "#/components/parameters/page" },
          { "$ref": "#/components/parameters/limit" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/FaqList" }
        }
      }
    },
    "/public/testimonials": {
      "get": {
        "tags": ["public/content"],
        "operationId": "listPublicTestimonials",
        "parameters": [
          { "$ref": "#/components/parameters/locale" },
          { "$ref": "#/components/parameters/page" },
          { "$ref": "#/components/parameters/limit" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/TestimonialList" }
        }
      }
    },
    "/public/gallery": {
      "get": {
        "tags": ["public/content"],
        "operationId": "listPublicGallery",
        "parameters": [
          { "$ref": "#/components/parameters/locale" },
          { "$ref": "#/components/parameters/page" },
          { "$ref": "#/components/parameters/limit" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/GalleryList" }
        }
      }
    },
    "/public/contact": {
      "post": {
        "tags": ["public/forms"],
        "operationId": "submitContact",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ContactRequest" }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/AcceptedMessage" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/public/custom-cake-requests": {
      "post": {
        "tags": ["public/forms"],
        "operationId": "submitCustomCakeRequest",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CustomCakeRequestInput" }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/AcceptedMessage" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
    "/admin/media/signature": {
      "post": {
        "tags": ["admin/media"],
        "operationId": "createMediaSignature",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MediaSignatureRequest" } } } },
        "responses": { "200": { "$ref": "#/components/responses/MediaSignature" } }
      }
    },
    "/admin/media/confirm": {
      "post": {
        "tags": ["admin/media"],
        "operationId": "confirmMedia",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MediaAssetInput" } } } },
        "responses": { "201": { "$ref": "#/components/responses/MediaItem" }, "400": { "$ref": "#/components/responses/ValidationError" } }
      }
    },
    "/admin/media/{id}": {
      "delete": {
        "tags": ["admin/media"],
        "operationId": "deleteMedia",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "responses": { "204": { "description": "Media eliminado sin cuerpo." }, "409": { "$ref": "#/components/responses/MediaInUse" } }
      }
    },
    "/admin/custom-cake-requests": {
      "get": {
        "tags": ["admin/custom-cake-requests"],
        "operationId": "listCustomCakeRequests",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/page" }, { "$ref": "#/components/parameters/limit" }],
        "responses": { "200": { "$ref": "#/components/responses/CustomCakeRequestList" } }
      }
    },
    "/admin/custom-cake-requests/{id}": {
      "get": {
        "tags": ["admin/custom-cake-requests"],
        "operationId": "getCustomCakeRequest",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "responses": { "200": { "$ref": "#/components/responses/CustomCakeRequestItem" }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "patch": {
        "tags": ["admin/custom-cake-requests"],
        "operationId": "updateCustomCakeRequest",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CustomCakeRequestPatch" } } } },
        "responses": { "200": { "$ref": "#/components/responses/CustomCakeRequestItem" } }
      }
    },
    "/admin/settings": {
      "get": {
        "tags": ["admin/settings"],
        "operationId": "getAdminSettings",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/SettingsItem" } }
      },
      "patch": {
        "tags": ["admin/settings"],
        "operationId": "updateAdminSettings",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SettingsPatch" } } } },
        "responses": { "200": { "$ref": "#/components/responses/SettingsItem" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "parameters": {
      "id": { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^[a-fA-F0-9]{24}$" } },
...
      "locale": { "name": "locale", "in": "query", "required": true, "schema": { "$ref": "#/components/schemas/Locale" } },
      "page": { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 10000, "default": 1 } },
      "limit": { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 24 } },
...
    },
...
    "responses": {
      "ValidationError": { "description": "Datos invalidos.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
...
      "Forbidden": { "description": "Sin permiso.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
...
      "NotFound": { "description": "No encontrado.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
...
      "RateLimited": { "description": "Rate limit.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
...
      "MediaInUse": { "description": "Media en uso.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "AcceptedMessage": { "description": "Mensaje aceptado.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageEnvelope" } } } },
...
      "FaqList": { "description": "FAQs.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FaqListEnvelope" } } } },
      "TestimonialList": { "description": "Testimonios.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TestimonialListEnvelope" } } } },
      "GalleryList": { "description": "Galeria.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GalleryListEnvelope" } } } },
      "MediaSignature": { "description": "Firma Cloudinary.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MediaSignatureEnvelope" } } } },
      "MediaItem": { "description": "Media.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MediaEnvelope" } } } },
      "CustomCakeRequestItem": { "description": "Solicitud.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CustomCakeRequestEnvelope" } } } },
      "CustomCakeRequestList": { "description": "Solicitudes.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CustomCakeRequestListEnvelope" } } } },
      "SettingsItem": { "description": "Configuracion.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SettingsEnvelope" } } } },
...
    },
    "schemas": {
      "Locale": { "type": "string", "enum": ["es", "en"] },
...
      "ErrorCode": { "type": "string", "enum": ["VALIDATION_ERROR", "UNAUTHORIZED", "FORBIDDEN", "INVALID_CREDENTIALS", "ACCOUNT_LOCKED", "TOKEN_EXPIRED", "TOKEN_INVALID", "EMAIL_ALREADY_EXISTS", "RESOURCE_NOT_FOUND", "SLUG_ALREADY_EXISTS", "SKU_ALREADY_EXISTS", "CATEGORY_NOT_EMPTY", "MEDIA_IN_USE", "CONFLICT", "RATE_LIMITED", "INTERNAL_ERROR"] },
      "Meta": {
        "type": "object",
        "properties": {
          "requestId": { "type": "string", "format": "uuid" },
          "timestamp": { "type": "string", "format": "date-time" }
        },
        "required": ["requestId", "timestamp"],
        "additionalProperties": false
      },
      "PaginationMeta": {
        "allOf": [
          { "$ref": "#/components/schemas/Meta" },
          {
            "type": "object",
            "properties": {
              "page": { "type": "integer", "minimum": 1 },
              "limit": { "type": "integer", "minimum": 1 },
              "totalItems": { "type": "integer", "minimum": 0 },
              "totalPages": { "type": "integer", "minimum": 1 }
            },
            "required": ["page", "limit", "totalItems", "totalPages"]
          }
        ]
      },
      "SuccessEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "const": true },
          "data": {},
          "meta": { "$ref": "#/components/schemas/Meta" }
        },
        "required": ["success", "data", "meta"],
        "additionalProperties": false
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "const": false },
          "error": {
            "type": "object",
            "properties": {
              "code": { "$ref": "#/components/schemas/ErrorCode" },
              "message": { "type": "string" },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": { "type": "string" },
                    "rule": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["field", "rule", "message"],
                  "additionalProperties": false
                }
              }
            },
            "required": ["code", "message"],
            "additionalProperties": false
          },
          "meta": { "$ref": "#/components/schemas/Meta" }
        },
        "required": ["success", "error", "meta"],
        "additionalProperties": false
      },
      "LocalizedText": {
        "type": "object",
        "properties": {
          "es": { "type": "string", "minLength": 1, "maxLength": 10000 },
          "en": { "type": "string", "minLength": 1, "maxLength": 10000 }
        },
        "required": ["es", "en"],
        "additionalProperties": false
      },
...
      "BrandColors": {
        "type": "object",
        "properties": {
          "brown": { "type": "string", "const": "#4A3C31" },
          "gold": { "type": "string", "const": "#D4AF37" },
          "goldSecondary": { "type": "string", "const": "#C5A059" },
          "cream": { "type": "string", "const": "#FDFBF7" }
        },
        "required": ["brown", "gold", "goldSecondary", "cream"],
        "additionalProperties": false
      },
      "PublicSettings": {
        "type": "object",
        "properties": {
          "brand": {
            "type": "object",
            "properties": { "colors": { "$ref": "#/components/schemas/BrandColors" } },
            "required": ["colors"],
            "additionalProperties": false
          }
        },
        "required": ["brand"],
        "additionalProperties": false
      },
      "MediaAsset": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "publicId": { "type": "string", "pattern": "^el-trigal/" },
          "secureUrl": { "type": "string", "format": "uri" },
          "resourceType": { "type": "string", "enum": ["image", "video"] },
          "format": { "type": "string" },
          "width": { "type": "integer", "minimum": 1, "maximum": 12000 },
          "height": { "type": "integer", "minimum": 1, "maximum": 12000 },
          "bytes": { "type": "integer", "minimum": 1 },
          "alt": { "$ref": "#/components/schemas/LocalizedText" },
          "caption": { "anyOf": [{ "$ref": "#/components/schemas/LocalizedText" }, { "type": "null" }] },
          "isPrimary": { "type": "boolean" },
          "sortOrder": { "type": "integer", "minimum": 0, "maximum": 10000 }
        },
        "required": ["id", "publicId", "secureUrl", "resourceType", "format", "width", "height", "bytes", "alt", "isPrimary", "sortOrder"],
        "additionalProperties": false
      },
...
      "Faq": {
        "type": "object",
        "properties": { "id": { "type": "string" }, "question": { "$ref": "#/components/schemas/LocalizedText" }, "answer": { "$ref": "#/components/schemas/LocalizedText" }, "sortOrder": { "type": "integer" } },
        "required": ["id", "question", "answer", "sortOrder"]
      },
      "Testimonial": {
        "type": "object",
        "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "quote": { "$ref": "#/components/schemas/LocalizedText" }, "rating": { "type": "integer", "minimum": 1, "maximum": 5 }, "sortOrder": { "type": "integer" } },
        "required": ["id", "name", "quote", "rating", "sortOrder"]
      },
      "GalleryItem": {
        "type": "object",
        "properties": { "id": { "type": "string" }, "title": { "$ref": "#/components/schemas/LocalizedText" }, "description": { "anyOf": [{ "$ref": "#/components/schemas/LocalizedText" }, { "type": "null" }] }, "media": { "$ref": "#/components/schemas/MediaAsset" }, "featured": { "type": "boolean" }, "sortOrder": { "type": "integer" } },
        "required": ["id", "title", "description", "media", "featured", "sortOrder"]
      },
      "ContactRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 2, "maxLength": 80 },
          "email": { "type": "string", "format": "email", "maxLength": 254 },
          "phone": { "type": ["string", "null"], "pattern": "^\\+[1-9]\\d{6,14}$" },
          "subject": { "type": "string", "minLength": 3, "maxLength": 100 },
          "message": { "type": "string", "minLength": 10, "maxLength": 2000 },
          "locale": { "$ref": "#/components/schemas/Locale" },
          "honeypot": { "type": "string", "maxLength": 0 }
        },
        "required": ["name", "email", "subject", "message", "locale"],
        "additionalProperties": false
      },
      "CustomCakeRequestInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 2, "maxLength": 80 },
          "email": { "type": "string", "format": "email", "maxLength": 254 },
          "phone": { "type": "string", "pattern": "^\\+[1-9]\\d{6,14}$" },
          "eventDate": { "type": "string", "format": "date" },
          "eventType": { "type": "string", "enum": ["birthday", "wedding", "quinceanera", "corporate", "other"] },
          "servings": { "type": "integer", "minimum": 1, "maximum": 500 },
          "budgetCents": { "type": "integer", "minimum": 0, "maximum": 1000000 },
          "flavor": { "type": "string", "minLength": 1, "maxLength": 80 },
          "filling": { "type": "string", "minLength": 1, "maxLength": 80 },
          "theme": { "type": "string", "minLength": 1, "maxLength": 120 },
          "message": { "type": "string", "maxLength": 2000 },
          "imageIds": { "type": "array", "maxItems": 5, "uniqueItems": true, "items": { "type": "string", "pattern": "^[a-fA-F0-9]{24}$" } },
          "locale": { "$ref": "#/components/schemas/Locale" },
          "consent": { "type": "boolean", "const": true },
          "honeypot": { "type": "string", "maxLength": 0 }
        },
        "required": ["name", "email", "phone", "eventDate", "eventType", "servings", "flavor", "filling", "theme", "locale", "consent"],
        "additionalProperties": false
      },
      "CustomCakeRequest": {
        "allOf": [
          { "$ref": "#/components/schemas/CustomCakeRequestInput" },
          {
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "images": { "type": "array", "items": { "$ref": "#/components/schemas/MediaAsset" } },
              "status": { "type": "string", "enum": ["new", "in-review", "quoted", "won", "lost", "archived"] },
              "internalNotes": { "type": ["string", "null"], "maxLength": 2000 },
              "createdAt": { "type": "string", "format": "date-time" },
              "updatedAt": { "type": "string", "format": "date-time" }
            }
          }
        ]
      },
      "CustomCakeRequestPatch": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["new", "in-review", "quoted", "won", "lost", "archived"] },
          "internalNotes": { "type": ["string", "null"], "maxLength": 2000 }
        },
        "additionalProperties": false
      },
...
      "MediaSignatureRequest": {
        "type": "object",
        "properties": { "folder": { "type": "string", "enum": ["el-trigal/products", "el-trigal/categories", "el-trigal/custom-cakes", "el-trigal/gallery", "el-trigal/business", "el-trigal/temporary"] }, "resourceType": { "type": "string", "enum": ["image", "video"] } },
        "required": ["folder", "resourceType"],
        "additionalProperties": false
      },
      "MediaSignatureData": {
        "type": "object",
        "properties": { "cloudName": { "type": "string" }, "apiKey": { "type": "string" }, "timestamp": { "type": "integer" }, "folder": { "type": "string" }, "resourceType": { "type": "string" }, "signature": { "type": "string" } },
        "required": ["cloudName", "apiKey", "timestamp", "folder", "resourceType", "signature"],
        "additionalProperties": false
      },
      "MediaAssetInput": {
        "allOf": [
          { "$ref": "#/components/schemas/MediaAsset" },
          {
            "type": "object",
            "properties": { "folder": { "type": "string", "enum": ["el-trigal/products", "el-trigal/categories", "el-trigal/custom-cakes", "el-trigal/gallery", "el-trigal/business", "el-trigal/temporary"] } },
            "required": ["folder"]
          }
        ]
      },
      "SettingsPatch": {
        "type": "object",
        "properties": {
          "brand": {
            "type": "object",
            "properties": {
              "colors": {
                "type": "object",
                "properties": {
                  "brown": { "type": "string", "pattern": "^#[0-9A-Fa-f]{6}$" },
                  "gold": { "type": "string", "pattern": "^#[0-9A-Fa-f]{6}$" },
                  "goldSecondary": { "type": "string", "pattern": "^#[0-9A-Fa-f]{6}$" },
                  "cream": { "type": "string", "pattern": "^#[0-9A-Fa-f]{6}$" }
                },
                "required": ["brown", "gold", "goldSecondary", "cream"],
                "additionalProperties": false
              }
            },
            "additionalProperties": false
          },
          "businessName": { "$ref": "#/components/schemas/LocalizedText" },
          "publicEmail": { "type": ["string", "null"], "format": "email", "maxLength": 254 },
          "publicPhone": { "type": ["string", "null"], "maxLength": 30 },
          "address": { "type": ["string", "null"], "maxLength": 200 }
        },
        "additionalProperties": false
      },
...
      "MessageEnvelope": {
        "allOf": [{ "$ref": "#/components/schemas/SuccessEnvelope" }, { "type": "object", "properties": { "data": { "type": "object", "properties": { "id": { "type": "string" }, "message": { "type": "string" }, "updated": { "type": "integer" } } } } }]
      },
...
      "FaqListEnvelope": { "allOf": [{ "$ref": "#/components/schemas/SuccessEnvelope" }, { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Faq" } }, "meta": { "$ref": "#/components/schemas/PaginationMeta" } } }] },
      "TestimonialListEnvelope": { "allOf": [{ "$ref": "#/components/schemas/SuccessEnvelope" }, { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Testimonial" } }, "meta": { "$ref": "#/components/schemas/PaginationMeta" } } }] },
      "GalleryListEnvelope": { "allOf": [{ "$ref": "#/components/schemas/SuccessEnvelope" }, { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/GalleryItem" } }, "meta": { "$ref": "#/components/schemas/PaginationMeta" } } }] },
      "MediaSignatureEnvelope": { "allOf": [{ "$ref": "#/components/schemas/SuccessEnvelope" }, { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/MediaSignatureData" } } }] },
      "MediaEnvelope": { "allOf": [{ "$ref": "#/components/schemas/SuccessEnvelope" }, { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/MediaAsset" } } }] },
      "CustomCakeRequestEnvelope": { "allOf": [{ "$ref": "#/components/schemas/SuccessEnvelope" }, { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/CustomCakeRequest" } } }] },
      "CustomCakeRequestListEnvelope": { "allOf": [{ "$ref": "#/components/schemas/SuccessEnvelope" }, { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/CustomCakeRequest" } }, "meta": { "$ref": "#/components/schemas/PaginationMeta" } } }] },
      "SettingsEnvelope": { "allOf": [{ "$ref": "#/components/schemas/SuccessEnvelope" }, { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/PublicSettings" } } }] },
...
    }
  }
}